+++ Verizon/CHANGELOG.md
@@ -1,10 +1,15 @@
 ## Changelog
+* 1.6.0.0
+    * Implemented Advanced Bidding token compression and encoding.
+    * Added Super Auction and Advanced Bidding to Rewarded Video. 
+    * This version of the adapters has been certified with Verizon 1.6.0.
+
 * 1.5.0.0
     * Compress the Advanced Bidding token to adhere to MoPub's spec.
     * This version of the adapters has been certified with Verizon 1.5.0 and MoPub 5.11.0.
 
 * 1.4.0.0
-    * Update Advanced Bidding API
+    * Updated Advanced Bidding API.
     * This version of the adapters has been certified with Verizon 1.4.0.
     
 * 1.3.1.0